requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LINE
line-bot-sdk==3.6.0

# AI/ML
anthropic==0.18.1
openai==1.12.0
google-cloud-vision==3.5.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Excel
openpyxl==3.1.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
